Improve readability of constructor calls of Riegeli classes.
Write the type name instead of `std::forward_as_tuple` for construction of
dependencies of these classes. This looks better, at the cost of a tiny
performance hit: the dependency is constructed as a temporary and then moved,
instead of being constructed in place.

Rely on CTAD
(https://en.cppreference.com/w/cpp/language/class_template_argument_deduction)
for Riegeli classes to avoid writing the class template argument, in cases the
template argument reflects the type of the first constructor argument. Riegeli
classes provide deduction guides to make it clear that this is supported.

CTAD requires C++17.

PiperOrigin-RevId: 364632707